Workaround for VS2008 compiler issue
When building in Release or RelWithDebInfo mode the code crashes when trying to perform a /= 100.0 on a float.
Explicitly casting the 100.0 to a float seems to solve the issue.

+++ libharp/harp-ingest-s5p_l2.c
@@ -3745,7 +3745,7 @@ static int read_sea_ice_fraction(void *user_data, harp_array data)
     {
         if (data.float_data[i] > 0 && data.float_data[i] <= 100)
         {
-            data.float_data[i] /= 100.0;
+            data.float_data[i] /= (float)100.0;
         }
         else
         {
